+++ sdk/textanalytics/Azure.AI.TextAnalytics/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Release History
 
 ## 5.1.0-beta.8 (Unreleased)
+### New features
+- Added value `None` to enum `PiiEntityDomainType` to allow user to specify no domain.
+
 ### Breaking changes
 - Renamed `StartAnalyzeBatchActions` to `StartAnalyzeActions`.
 - Renamed `AnalyzeBatchActionsOperation` to `AnalyzeActionsOperation`.
@@ -13,6 +16,8 @@
   - `RecognizeLinkedEntitiesOptions` changed to new type `RecognizeLinkedEntitiesActions`.
   - `AnalyzeSentimentOptions` changed to new type `AnalyzeSentimentActions`.
 - Renamed type `TextAnalyticsActionDetails` to `TextAnalyticsActionResult`.
+- Changed type `RecognizePiiEntitiesOptions.DomainFilter` from `PiiEntityDomainType?` to `PiiEntityDomainType`.
+- Renamed type `PiiEntityDomainType` to `PiiEntityDomain`.
 
 ## 5.1.0-beta.7 (2021-05-18)
 ### New features